use sandbox event manager even when a installed exist

+++ Makefile
@@ -84,7 +84,7 @@ coverage-event-manager: force
 sandbox: misc/env.sh
 	mkdir -p sandbox/${PREFIX}/lib64
 	cp -p misc/env.sh sandbox/env.sh
-	ln -s lib64 sandbox/${PREFIX}/lib
+	test -e sandbox/${PREFIX}/lib || ln -s lib64 sandbox/${PREFIX}/lib
 
 test-uzbl-core: uzbl-core
 	./uzbl-core --uri http://www.uzbl.org --verbose
@@ -98,16 +98,16 @@ test-uzbl-core-sandbox: sandbox uzbl-core sandbox-install-uzbl-core sandbox-inst
 	rm -rf ./sandbox/usr
 
 test-uzbl-browser-sandbox: sandbox uzbl-browser sandbox-install-uzbl-browser sandbox-install-example-data
-	-. ./sandbox/env.sh && uzbl-event-manager restart -avv
+	. ./sandbox/env.sh && python -S `which uzbl-event-manager` restart -avv
 	. ./sandbox/env.sh && uzbl-browser --uri http://www.uzbl.org --verbose
-	. ./sandbox/env.sh && uzbl-event-manager stop -vv -o /dev/null
+	. ./sandbox/env.sh && python -S `which uzbl-event-manager` stop -vv -o /dev/null
 	make DESTDIR=./sandbox uninstall
 	rm -rf ./sandbox/usr
 
 test-uzbl-tabbed-sandbox: sandbox uzbl-browser sandbox-install-uzbl-browser sandbox-install-uzbl-tabbed sandbox-install-example-data
-	-. ./sandbox/env.sh && uzbl-event-manager restart -avv
+	. ./sandbox/env.sh && python -S `which uzbl-event-manager` restart -avv
 	. ./sandbox/env.sh && uzbl-tabbed
-	. ./sandbox/env.sh && uzbl-event-manager stop -ivv
+	. ./sandbox/env.sh && python -S `which uzbl-event-manager` stop -ivv
 	make DESTDIR=./sandbox uninstall
 	rm -rf ./sandbox/usr
